ocr_project/requirements.txt
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.26.1
# Advanced image processing dependencies
scipy==1.11.3
scikit-learn==1.3.1
scikit-image==0.21.0
# Document processing dependencies
python-docx==0.8.11
pandas==2.1.3
PyMuPDF==1.23.8
pdfplumber==0.9.0
PyPDF2==3.0.1
lxml==4.9.3
# web interface
flask==3.0.0
# database
pymongo==4.15.4
dnspython==2.8.0
# deep learning ocr
easyocr==1.7.0
paddlepaddle==2.5.1
paddleocr==2.7.0.3
torch==2.0.1
torchvision==0.15.2
# dev
pytest==7.4.0
